[AIRFLOW-6721] Move HiveToMySqlTransfer test
Move the HiveToMySqlTransfer test in the operators module into
the existing test module and use the shared test environment for
these tests.
